requirements.txt
pytorch
transformers
textacy